add Beartrap (Trapper)
