+++ Roles/AddOns/Common/Beartrap.cs
@@ -1,0 +1,43 @@
+﻿using System.Collections.Generic;
+using static TOHE.Options;
+
+namespace TOHE.Roles.AddOns.Common
+{
+    public static class Trapper
+    {
+        private static readonly int Id = 18800;
+
+
+        public static OptionItem ImpCanBeTrapper;
+        public static OptionItem CrewCanBeTrapper;
+        public static OptionItem NeutralCanBeTrapper;
+        public static OptionItem TrapperBlockMoveTime;
+
+
+        public static void SetupCustomOptions()
+        {
+            SetupAdtRoleOptions(Id, CustomRoles.Trapper, canSetNum: true);
+            ImpCanBeTrapper = BooleanOptionItem.Create(Id + 10, "ImpCanBeTrapper", true, TabGroup.Addons, false).SetParent(CustomRoleSpawnChances[CustomRoles.Trapper]);
+            CrewCanBeTrapper = BooleanOptionItem.Create(Id + 11, "CrewCanBeTrapper", true, TabGroup.Addons, false).SetParent(CustomRoleSpawnChances[CustomRoles.Trapper]);
+            NeutralCanBeTrapper = BooleanOptionItem.Create(Id + 12, "NeutralCanBeTrapper", true, TabGroup.Addons, false).SetParent(CustomRoleSpawnChances[CustomRoles.Trapper]);
+            TrapperBlockMoveTime = FloatOptionItem.Create(Id + 13, "TrapperBlockMoveTime", new(1f, 180f, 1f), 5f, TabGroup.Addons, false).SetParent(CustomRoleSpawnChances[CustomRoles.Trapper])
+                .SetValueFormat(OptionFormat.Seconds);
+        }
+        public static void TrapperKilled(this PlayerControl killer, PlayerControl target)
+        {
+            Logger.Info($"{target?.Data?.PlayerName} was Trapper", "Trapper");
+            var tmpSpeed = Main.AllPlayerSpeed[killer.PlayerId];
+            Main.AllPlayerSpeed[killer.PlayerId] = Main.MinSpeed;
+            ReportDeadBodyPatch.CanReport[killer.PlayerId] = false;
+            killer.MarkDirtySettings();
+            _ = new LateTask(() =>
+            {
+                Main.AllPlayerSpeed[killer.PlayerId] = Main.AllPlayerSpeed[killer.PlayerId] - Main.MinSpeed + tmpSpeed;
+                ReportDeadBodyPatch.CanReport[killer.PlayerId] = true;
+                killer.MarkDirtySettings();
+                RPC.PlaySoundRPC(killer.PlayerId, Sounds.TaskComplete);
+            }, Trapper.TrapperBlockMoveTime.GetFloat(), "Trapper BlockMove");
+        }
+
+    }
+}
