+++ Roles/Impostor/Vampire.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using TOHE.Modules;
 using TOHE.Roles.Crewmate;
 using UnityEngine;
+using TOHE.Roles.AddOns.Common;
 using static TOHE.Translator;
 
 namespace TOHE.Roles.Impostor;
