+++ Roles/Neutral/Poisoner.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using TOHE.Roles.Crewmate;
 using UnityEngine;
+using TOHE.Roles.AddOns.Common;
 using static TOHE.Translator;
 
 namespace TOHE.Roles.Neutral;
